Move color processing to manga fetcher + own scope
